% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtw-package.R
\docType{package}
\name{obtw-package}
\alias{obtw}
\alias{obtw-package}
\title{obtw: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/krlmlr/duckplyr}
  \item \url{https://krlmlr.github.io/duckplyr}
  \item Report bugs at \url{https://github.com/krlmlr/duckplyr/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{kirill@cynkra.com} (\href{https://orcid.org/0000-0002-1416-3412}{ORCID})

}
\keyword{internal}
